refactor: use search bar inside store package
